Clean up catlin-linter and diff-task-job usage
This commit refactors the kustomization files for the `community` and `verified` catalogs:

- Run the `diff-task-job` only for the community catalog repo which uses the legacy directory-based versioning strategy, and skip it for the verified catalogs which uses git-based versioning strategy.
- Kustomize the GitHub Check name to be updated based on the catalog repo:
  - Use `pull-catalog-catlin-lint` Github Check for community catalog (no change)
  - Use `pull-tekton-catalog-build-tests` Github Check for verified catalog

This commit also fixes the bug in the `catlin-linter` pipeline. Prior to this change, the `lint-catalog` pipelineTask is NOT  guarded by the previous pipelineTask `check-name-match`' result value. When the `check-name-match` pipelineTask fails, the depedent `clone-repo` pipelineTask is skipped. However the order dependent `lint-catalog` pipelineTask is still executed and will always pass.  The success of `lint-catalog` pipelineTask emit cloud events to make the corresponding Github Check always pass.

This commit fixes the issue by adding a `when` expression to guard the `lint-catalog` pipelineTask.