Update Dashboard nightly tests on Z and P to use nightly releases
Instead of building the Dashboard from main, run the tests against
the latest nightly release of the Dashboard.

This yields significant performance improvements in the tests,
for example on prow it reduced from ~17-18 mins to ~8-9 mins. We
should see similar reduction on dogfooding, getting quicker feedback
and also freeing up the cluster for other work.