Remove PIPELINE_FEATURE_GATE env var
This environment variable is used for setting the correct value of
"enable-api-fields" for backports to Pipelines before the files
test/e2e-tests-kind-prow-<flag>.env were added. However, the earliest
version of Pipelines that is still supported is v0.41.0, which does have
the .env files. Therefore, this fallback is no longer needed. This commit
removes this logic to reduce confusion.