Unblocked fix-check-pr-has-kind-label CI
`fix-check-pr-has-kind-label` check started to fail due to [yaml/pyyaml#601]. This commit updates the `PyYaml` install command to use an older version of `Cython`,
which is a workaround to unblock our CI as suggested [here](https://github.com/yaml/pyyaml/issues/601#issuecomment-1638509577).

This should be a temporary fix. There is a long-term fix in [yaml/pyyaml#720].
Tested it locally and `PyYaml` seems to install and work properly.

/kind misc

[yaml/pyyaml#720]: https://github.com/yaml/pyyaml/pull/702
[yaml/pyyaml#601]: https://github.com/yaml/pyyaml/issues/601