Fix Rerun and Trigger Commands
Prior to this change, some rerun and trigger commands did not match
the name of the tests. This leads to confusion when trying to use
the commands on a pull request.

Ran into this issue when reruning https://github.com/tektoncd/pipeline/pull/6659.

This change adds the prefix `"pull-"` to commands that were missing
it, while the test name has that prefix.