Remove unused file from Pipelines nightly deploy
Robocat is stuck on Pipelines `v20230307-771fb665f2` due to an error
in the nightly deploy TaskRun.

See for example from the most recent run:
```
2024-02-15T03:01:00.091452354Z + kubectl apply --kubeconfig /workspace/targetCluster/kubeconfig -k pipeline/overlays/robocat
2024-02-15T03:01:02.736006696Z error: no matches for Id ~G_v1_ConfigMap|tekton-pipelines|config-artifact-bucket; failed to find unique target for patch ~G_v1_ConfigMap|config-artifact-bucket
```

The referenced ConfigMap was removed from Pipelines in https://github.com/tektoncd/pipeline/pull/6150 which was merged on March 7th.

Remove the file from the overlay as it's no longer used.