Rework plumbing nightlies
Define a reusable task and pipeline in plumbing under
tekton/catalog that can be used by all components in the
plumbing repo.

Move the five components in the repo to the shared
resources, and in doing that, point the new nightly
builds to ghcr.io and start skipping build and unit
tests at nightly build time.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>