Fix failing kind cluster start and update docs
When starting the kind cluster with the previous configuration the
control-plane kubelet errored out because it does not have a feature
gate for "EphemeralContainers". This feature is stable since kubernetes
1.24 so this configuration can safely be removed.

Added logs with info about successful and next steps which helps to
debug when something fails.

As of this writing the support for podman in kind is experimental. See
https://github.com/kubernetes-sigs/kind/issues/1778.

When podman and docker are both installed on the host the previous
script started the registry using podman but kind using docker. This is
fixed by changing the default container runtime to be docker. Docker
is also used in the `local-setup.md` script which is linked in the
official docs here
https://tekton.dev/docs/installation/local-installation.

Updated the documentation to reflect the changes. Added a section on
how to use podman for the registry and kind.

Fixed broken links in CONTRIBUTING.md. Updated cluster installation docs
in DEVELOPMENT.md to use latest kind and newer go versions, added
links to kustomize and tkn installation guides.

Issue #2486

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>