Fix kind cluster issues on Fedora and update docs
When starting the kind cluster using docker with the previous
configuration the control-plane kubelet stopped with an error not
knowing a feature gate for "EphemeralContainers". This feature is stable
since kubernetes 1.24 so this configuration was safely removed.

Added logs with info about successful and next steps which helps to
debug when something fails.

Starting the kind cluster using podman failed when joining the
worker-nodes. This is fixed by a minor change on the host. A
troubleshooting section was added to the readme with a link to the fix.

Once the kind cluster started using podman the registry could not be
connected. This issue was resolved by changing the network mode to
"bridge" which is also used in the guide on the kind website.

If podman and docker runtimes are installed on the host the script does
not behave as expected. Starting it using podman will start the registry
in podman and the kind cluster in docker. It was updated to always start
the registry and the cluster using the same runtime.

The documentation cleanup section was extended to cover the different
ways to delete the cluster depending on the runtime.

As of this writing the support for podman in kind is experimental. See
https://github.com/kubernetes-sigs/kind/issues/1778. This information
was added to the documentation.

Fixed broken links in CONTRIBUTING.md. Updated cluster installation docs
in DEVELOPMENT.md to use latest kind and newer go versions, added
links to kustomize and tkn installation guides.

Issues #2486, #2522, #2523

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>