Add kustomization file for cm
Add a kustomization file that can be used to deploy prow along
with its config maps. I'm replicating the prow setup on OCI,
using ArgoCD to sync folders from the plumbing repo.
This kustomization file allows to use the kustomize plugin
in ArgoCD and avoid the extra process to generate config maps.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>