Prepare for Tekton CI on OCI + cleanup
Start preparing tekton-ci resources for the new environment
on OCI. We need a new webhook URL, remove all the Gubernator
configurations and tasks (logs are stored in OCI buckets
and served via the Tekton dashboard), cleanup legacy boskos
configs and update documentation.

Tekton CI requires custom cluster interceptors and custom
task controllers to be deployed, so adding those into
tekton/cd for argocd to deploy.

The build-id doesn't have nightly builds for now, so it's
not installed yet, it will have to be added for things to work
once a build is available.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>